requirements.txt
keyboard==0.13.5
numpy==1.26.1
Pillow==10.1.0
pyautogui==0.9.54
python_imagesearch==1.3.0
